pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "observability-log-py"
version = "0.2.0"
description = "Python observability helpers for trace/log correlation"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "opentelemetry-api>=1.39.0",
]

[project.optional-dependencies]
langfuse = [
  "langfuse>=3.14.5",
  "opentelemetry-sdk>=1.39.0",
  "litellm>=1.80.7",
  "starlette>=0.40.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["obslogpy*"]
